move functions and errors from registry to common packages
